+++ src/main/java/racingcar/util/RandomUtils.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import camp.nextstep.edu.missionutils.Randoms;
