+++ src/main/java/racingcar/view/Input.java
@@ -1,9 +1,8 @@
-package racingcar;
+package racingcar.view;
 
 import camp.nextstep.edu.missionutils.Console;
 
-public class InputUtils {
-
+public class Input {
     public static String[] inputCarNames() {
         System.out.println("경주할 자동차 이름을 입력하세요. (이름은 쉼표(,) 기준으로 구분)");
         String inputStr = Console.readLine();
