+++ src/test/java/racingcar/util/RandomUtilsTest.java
@@ -1,8 +1,9 @@
-package racingcar;
+package racingcar.util;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.Test;
+import racingcar.util.RandomUtils;
 
 class RandomUtilsTest {
